requirements.txt
asgiref==3.8.1
awsebcli==3.21.0
blessed==1.20.0
botocore==1.35.99
cement==2.10.14
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
idna==3.10
jmespath==1.0.1
mysqlclient==2.2.7
pathspec==0.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
requests==2.32.3
semantic-version==2.10.0
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
termcolor==2.5.0
urllib3==1.26.20
wcwidth==0.2.13
